Undo the OS detection in favor of a weird macro
The OS detection and my Startech did not agree at all. Mostly I
had a non-working keyboard when plugged into the HID port, which is
the port that is needed if you want to use the keyboard-based
connection switching.

So this, instead, adds a macro and a bool to switch the mode, and
steals a key (KC_MPLY) that was mostly just opening Apple Music when I
didn't want it to, anyway.